Move WeightInfo, Config, Event into Portfolio pallet.
