+++ pallets/sto/src/lib.rs
@@ -38,7 +38,6 @@ use sp_std::prelude::*;
 use pallet_base::try_next_post;
 use pallet_identity::PermissionedCallOriginData;
 use pallet_settlement::VenueInfo;
-use polymesh_common_utilities::traits::portfolio;
 use polymesh_primitives::asset::AssetId;
 use polymesh_primitives::impl_checked_inc;
 use polymesh_primitives::settlement::{Leg, ReceiptDetails, SettlementType, VenueId, VenueType};
@@ -170,7 +169,7 @@ pub trait Config:
     frame_system::Config
     + pallet_identity::Config
     + pallet_settlement::Config
-    + portfolio::Config
+    + pallet_portfolio::Config
     + pallet_base::Config
 {
     /// The overarching event type.
